Created CubeAudio and MusicRoutine
CubeAudio is a class that simplifies the process of getting arrays of audio data.

MusicRoutine uses CubeAudio to turn the cube into a visualization of the amplitudes at different frequencies

Also included are two functions that should be added to LEDCube or CoordinateSystem, discretizecube and lightcollumn. These break up the x,y cube into squares, and light up a vertical string of LEDs, respectively